Test the "user"
To be used in class on review day.  You are to create a "User" class that works with this test file.  See instructions on haiku, or <a href="https://docs.google.com/a/hb.edu/document/d/1da7u35KW8BH2FBJobl_JagIkjL4FuhmLGiQfu06djbY/edit?usp=sharing">here</a>.